app/requirements.txt
asgiref==3.3.4
cffi==1.14.5
PyCryptodome==3.4.7
Django==3.2
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
phonenumbers==8.12.21
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
django-cryptography